pyproject.toml
[project]
name = "autobilanceo"
version = "0.1.0"
description = "Automated web navigation tooling for AFIP website, for dummies"
readme = "README.md"
authors = [{ name = "minollisantiago", email = "minollisantiago@gmail.com" }]
requires-python = ">=3.13"
dependencies = [
  "playwright>=1.42.0",
  "pandas>=2.2.0",
  "gspread>=6.0.0",
  "google-auth>=2.28.0",
  "python-dotenv>=1.0.0",
  "pydantic>=2.6.0",
  "rich>=13.7.0",
  "typer>=0.9.0",
]

[project.scripts]
test_auth = "autoBilanceo.tests.test_auth:run"
test_comp_nav = "autoBilanceo.tests.services.comprobantes.test_step1_nav_to_generator:run"
test_comp_type = "autoBilanceo.tests.services.comprobantes.test_step2_select_invoice_type:run"
test_comp_form_1 = "autoBilanceo.tests.services.comprobantes.test_step3_fill_invoice_issuance_form:run"
test_comp_form_2 = "autoBilanceo.tests.services.comprobantes.test_step4_fill_recipient_form:run"
test_comp_form_3 = "autoBilanceo.tests.services.comprobantes.test_step5_fill_invoice_content_form:run"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/autoBilanceo"]

[project.optional-dependencies]
dev = ["pytest>=8.0.0"]
